Tools\FitGen: FIX for GCC Build Failure
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2981

Added Type Casting to print UINT64 Value using Printf to fix the GCC build error.

Cc: Bob Feng <bob.c.feng@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Liu, Zhiguang <zhiguang.liu@intel.com>

+++ Silicon/Intel/Tools/FitGen/FitGen.c
@@ -1004,7 +1004,7 @@ Returns:
     //
     gFitTableContext.TopFlashAddressRemapValue = 0x100000000;
   }
-  printf ("Top Flash Address Value : 0x%llx\n", gFitTableContext.TopFlashAddressRemapValue);
+  printf ("Top Flash Address Value : 0x%llx\n", (unsigned long long) gFitTableContext.TopFlashAddressRemapValue);
   //
   // 0.4 Clear FIT table related memory
   //
